refactor: remove Quartz scheduler, handle medicine deletion errors, and restructure configurations
- Deprecate Quartz scheduler and related configuration in `application.yml`, `pom.xml`, and associated classes.
- Implement `CannotDeleteMedicine` exception and handling for cases with reminder logs.
- Adjust `MedicineService` and `ReminderService` by removing `SchedulerException`.
- Restructure configuration files and classes under the `shared` module for consistency.

+++ src/main/java/com/nexaworks/rafiq/medication/service/implementation/MedicineServiceImpl.java
@@ -8,7 +8,6 @@ import java.util.Optional;
 import java.util.UUID;
 
 import org.jetbrains.annotations.NotNull;
-import org.quartz.SchedulerException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -17,14 +16,11 @@ import org.springframework.transaction.annotation.Transactional;
 import com.nexaworks.rafiq.medication.api.dto.request.BulkMedicineOperationRequest;
 import com.nexaworks.rafiq.medication.api.dto.request.MedicineFilter;
 import com.nexaworks.rafiq.medication.api.dto.request.UpdateMedicinePatchRequest;
+import com.nexaworks.rafiq.medication.entity.enums.MedicineStatus;
 import com.nexaworks.rafiq.medication.entity.model.Drug;
 import com.nexaworks.rafiq.medication.entity.model.Group;
 import com.nexaworks.rafiq.medication.entity.model.Medicine;
-import com.nexaworks.rafiq.medication.entity.enums.MedicineStatus;
-import com.nexaworks.rafiq.medication.exception.GroupNotFoundException;
-import com.nexaworks.rafiq.medication.exception.MedicineAlreadyExist;
-import com.nexaworks.rafiq.medication.exception.MedicineLimit;
-import com.nexaworks.rafiq.medication.exception.MedicineNotFound;
+import com.nexaworks.rafiq.medication.exception.*;
 import com.nexaworks.rafiq.medication.repository.MedicineRepository;
 import com.nexaworks.rafiq.medication.repository.specification.MedicineSpecification;
 import com.nexaworks.rafiq.medication.service.DrugService;
@@ -45,7 +41,7 @@ public class MedicineServiceImpl implements MedicineService {
 
     @Override
     @Transactional
-    public Medicine addMedicine(Medicine entity, UUID drugId,UUID patientId) {
+    public Medicine addMedicine(Medicine entity, UUID drugId, UUID patientId) {
 
         entity.setPatientId(patientId);
         if (medicineRepository.existsByPatientIdAndDrugId(patientId, drugId)) {
@@ -63,7 +59,8 @@ public class MedicineServiceImpl implements MedicineService {
     }
 
     @Override
-    public Page<Medicine> getAllMedicines(Pageable pageable, MedicineFilter filter,UUID patientId) {
+    public Page<Medicine> getAllMedicines(Pageable pageable, MedicineFilter filter,
+            UUID patientId) {
 
         return medicineRepository.findAll(MedicineSpecification.filter(filter, patientId),
                 pageable);
@@ -71,20 +68,26 @@ public class MedicineServiceImpl implements MedicineService {
 
     @Override
     @Transactional(readOnly = true)
-    public Medicine getMedicineById(UUID medicineId,UUID patientId) {
+    public Medicine getMedicineById(UUID medicineId, UUID patientId) {
         return getMedicine(medicineId, patientId);
     }
 
     @Override
     @Transactional
-    public void deleteMedicine(UUID medicineId,UUID patientId) {
+    public void deleteMedicine(UUID medicineId, UUID patientId) {
         Medicine medicine = getMedicine(medicineId, patientId);
+        if (!medicine.getReminder().getReminderLogs().isEmpty()) {
+            medicine.setStatus(MedicineStatus.INACTIVE);
+            medicineRepository.save(medicine);
+            throw new CannotDeleteMedicine(
+                    "Medicine has reminder logs associated with it and cannot be deleted");
+        }
         medicineRepository.delete(medicine);
     }
 
     @Override
     @Transactional
-    public Medicine updateMedicine(Medicine entity, UUID medicineId,UUID patientId) {
+    public Medicine updateMedicine(Medicine entity, UUID medicineId, UUID patientId) {
         Medicine medicine = getMedicine(medicineId, patientId);
         medicine.setDosage(entity.getDosage());
         medicine.setFrequency(entity.getFrequency());
@@ -101,8 +104,8 @@ public class MedicineServiceImpl implements MedicineService {
 
     @Override
     @Transactional
-    public Medicine updateSpecific(UUID medicineId, UpdateMedicinePatchRequest request,UUID patientId)
-            throws SchedulerException {
+    public Medicine updateSpecific(UUID medicineId, UpdateMedicinePatchRequest request,
+            UUID patientId) {
         Medicine medicine = getMedicine(medicineId, patientId);
         request.dosage().ifPresent(medicine::setDosage);
         request.frequency().ifPresent(medicine::setFrequency);
@@ -129,23 +132,24 @@ public class MedicineServiceImpl implements MedicineService {
 
     @Override
     @Transactional
-    public List<UUID> bulkMedicineOperation(BulkMedicineOperationRequest request,UUID patientId)
+    public List<UUID> bulkMedicineOperation(BulkMedicineOperationRequest request, UUID patientId)
             throws NoSuchMethodException, InvocationTargetException, IllegalAccessException {
         if (request.medicineIds().isEmpty()) {
             throw new ValidationException("Medicine ids cannot be empty");
         }
         List<UUID> failedIds = new ArrayList<>();
         Method action = this.getClass().getMethod(request.action().getAction(), List.class,
-                Optional.class, List.class,UUID.class);
-        action.invoke(this, request.medicineIds(), request.groupId(), failedIds,patientId);
+                Optional.class, List.class, UUID.class);
+        action.invoke(this, request.medicineIds(), request.groupId(), failedIds, patientId);
 
         return failedIds;
     }
     @Transactional
-    public void delete(List<UUID> ids, Optional<UUID> groupId, List<UUID> failedIds,UUID patientId) {
+    public void delete(List<UUID> ids, Optional<UUID> groupId, List<UUID> failedIds,
+            UUID patientId) {
         ids.forEach(medicineId -> {
             try {
-                deleteMedicine(medicineId,patientId);
+                deleteMedicine(medicineId, patientId);
             } catch (Exception e) {
                 failedIds.add(medicineId);
             }
@@ -153,7 +157,8 @@ public class MedicineServiceImpl implements MedicineService {
 
     }
     @Transactional
-    public void moveToGroup(List<UUID> ids, Optional<UUID> groupId, List<UUID> failedIds,UUID patientId) {
+    public void moveToGroup(List<UUID> ids, Optional<UUID> groupId, List<UUID> failedIds,
+            UUID patientId) {
         groupId.orElseThrow(() -> new GroupNotFoundException("Group id is required"));
         Group group = groupService.getGroupById(groupId.get(), patientId);
         if (!group.getPatientId().equals(patientId)) {
@@ -170,7 +175,8 @@ public class MedicineServiceImpl implements MedicineService {
         });
     }
     @Transactional
-    public void markActive(List<UUID> ids, Optional<UUID> groupId, List<UUID> failedIds,UUID patientId) {
+    public void markActive(List<UUID> ids, Optional<UUID> groupId, List<UUID> failedIds,
+            UUID patientId) {
         ids.forEach(medicineId -> {
             try {
                 Medicine medicine = getMedicine(medicineId, patientId);
@@ -183,7 +189,8 @@ public class MedicineServiceImpl implements MedicineService {
 
     }
     @Transactional
-    public void markInActive(List<UUID> ids, Optional<UUID> groupId, List<UUID> failedIds,UUID patientId) {
+    public void markInActive(List<UUID> ids, Optional<UUID> groupId, List<UUID> failedIds,
+            UUID patientId) {
         ids.forEach(medicineId -> {
             try {
                 Medicine medicine = getMedicine(medicineId, patientId);
